[RHELC-1337] Port checks.check_kernel_boot_files() to Action framework
